refactor: new file for dummy handler
